split test_common
