+++ tests/test_common_select.py
@@ -1,0 +1,442 @@
+from __future__ import annotations
+
+import pytest
+
+import pandas as pd
+
+from io import StringIO
+
+
+from cdm_reader_mapper.common.select import (
+    _split_df,
+    _split_by_index_df,
+    _split_by_boolean_df,
+    _split_by_column_df,
+    split_by_boolean,
+    split_by_boolean_true,
+    split_by_boolean_false,
+    split_by_column_entries,
+    split_by_index,
+)
+
+
+def make_parser(text, **kwargs):
+    """Helper: create a TextFileReader similar to user code."""
+    buffer = StringIO(text)
+    return pd.read_csv(buffer, chunksize=2, **kwargs)
+
+
+def dummy_func(x):
+    return 2 * x
+
+
+@pytest.fixture
+def sample_df():
+    return pd.DataFrame(
+        {
+            "A": [1, 2, 3, 4, 5],
+            "B": ["x", "y", "z", "x", "y"],
+            "C": [True, False, True, False, True],
+        },
+        index=[10, 11, 12, 13, 14],
+    )
+
+
+@pytest.fixture
+def sample_reader():
+    text = "10,1,x,True\n11,2,y,False\n12,3,z,True\n13,4,x,False\n14,5,y,True"
+    reader = make_parser(text, names=["A", "B", "C"])
+    return reader
+
+
+@pytest.fixture
+def sample_df_multi():
+    return pd.DataFrame(
+        {
+            "A": [1, 2, 3, 4, 5],
+            "B": ["x", "y", "z", "x", "y"],
+            "C": [True, False, True, False, True],
+        },
+        index=[10, 11, 12, 13, 14],
+    )
+
+
+@pytest.fixture
+def sample_reader_multi():
+    text = "10,1,x,True\n11,2,y,False\n12,3,z,True\n13,4,x,False\n14,5,y,True"
+    reader = make_parser(text, names=[("A", "a"), ("B", "b"), ("C", "c")])
+    return reader
+
+
+@pytest.fixture
+def empty_df():
+    return pd.DataFrame(columns=["A", "B", "C"])
+
+
+@pytest.fixture
+def empty_reader():
+    return make_parser("", names=["A", "B", "C"])
+
+
+@pytest.fixture
+def boolean_mask():
+    return pd.DataFrame(
+        {
+            "mask1": [False, True, False, True, False],
+            "mask2": [True, False, True, False, True],
+        },
+        index=[10, 11, 12, 13, 14],
+    )
+
+
+@pytest.fixture
+def boolean_mask_true():
+    return pd.DataFrame(
+        {
+            "mask1": [True, True, False, True, False],
+            "mask2": [True, False, True, False, True],
+        },
+        index=[10, 11, 12, 13, 14],
+    )
+
+
+def test_split_df(sample_df):
+    mask = pd.Series([True, False, False, True, False], index=sample_df.index)
+    selected, rejected, _, _ = _split_df(sample_df, mask, return_rejected=True)
+    assert list(selected.index) == [10, 13]
+    assert list(rejected.index) == [11, 12, 14]
+
+
+def _test_split_df_false_mask(sample_df):
+    mask = pd.Series([False, False, False, False, False], index=sample_df.index)
+    selected, rejected, _, _ = _split_df(sample_df, mask, return_rejected=True)
+    assert list(selected.index) == [10, 13]
+    assert list(rejected.index) == [11, 12, 14]
+
+
+def test_split_df_multiindex(sample_df):
+    mask = pd.Series([True, False, False, True, False], index=sample_df.index)
+    sample_df.columns = pd.MultiIndex.from_tuples(
+        [
+            ("A", "a"),
+            (
+                "B",
+                "b",
+            ),
+            ("C", "c"),
+        ]
+    )
+    selected, rejected, _, _ = _split_df(sample_df, mask, return_rejected=True)
+    assert list(selected.index) == [10, 13]
+    assert list(rejected.index) == [11, 12, 14]
+
+
+@pytest.mark.parametrize(
+    "column,boolean,expected_selected,expected_rejected",
+    [
+        ("C", True, [10, 12, 14], [11, 13]),
+        ("C", False, [11, 13], [10, 12, 14]),
+    ],
+)
+def test_split_by_boolean_df(
+    sample_df, column, boolean, expected_selected, expected_rejected
+):
+    mask = sample_df[[column]]
+    selected, rejected, _, _ = _split_by_boolean_df(
+        sample_df, mask, boolean=boolean, return_rejected=True
+    )
+    assert list(selected.index) == expected_selected
+    assert list(rejected.index) == expected_rejected
+
+
+def test_split_by_boolean_df_empty_mask(sample_df):
+    mask = pd.DataFrame(columns=sample_df.columns)
+    selected, rejected, _, _ = _split_by_boolean_df(
+        sample_df, mask, boolean=True, return_rejected=True
+    )
+    assert list(selected.index) == list(sample_df.index)
+    assert rejected.empty
+
+
+@pytest.mark.parametrize(
+    "col,values,return_rejected,expected_selected,expected_rejected",
+    [
+        ("B", ["x", "z"], True, [10, 12, 13], [11, 14]),
+        ("B", ["missing"], True, [], [10, 11, 12, 13, 14]),
+        ("B", ["x", "z"], False, [10, 12, 13], []),
+    ],
+)
+def test_split_by_column_df(
+    sample_df, col, values, return_rejected, expected_selected, expected_rejected
+):
+    selected, rejected, _, _ = _split_by_column_df(
+        sample_df, col, values, return_rejected=return_rejected
+    )
+    assert list(selected.index) == expected_selected
+    assert list(rejected.index) == expected_rejected
+
+
+@pytest.mark.parametrize(
+    "index_list,inverse,return_rejected,expected_selected,expected_rejected",
+    [
+        ([11, 13], False, True, [11, 13], [10, 12, 14]),
+        ([11, 13], False, False, [11, 13], []),
+        ([11, 13], True, True, [10, 12, 14], [11, 13]),
+    ],
+)
+def test_split_by_index_df(
+    sample_df,
+    index_list,
+    inverse,
+    return_rejected,
+    expected_selected,
+    expected_rejected,
+):
+    selected, rejected, _, _ = _split_by_index_df(
+        sample_df, index_list, inverse=inverse, return_rejected=return_rejected
+    )
+    assert list(selected.index) == expected_selected
+    assert list(rejected.index) == expected_rejected
+
+
+@pytest.mark.parametrize("TextFileReader", [False, True])
+def test_split_by_index_basic(sample_df, sample_reader, TextFileReader):
+    if TextFileReader:
+        data = sample_reader
+    else:
+        data = sample_df
+    selected, rejected, _, _ = split_by_index(data, [11, 13], return_rejected=True)
+
+    if TextFileReader:
+        selected = selected.read()
+        rejected = rejected.read()
+
+    assert list(selected.index) == [11, 13]
+    assert list(rejected.index) == [10, 12, 14]
+
+
+def test_split_by_index_multiindex(sample_reader_multi):
+    selected, rejected, _, _ = split_by_index(
+        sample_reader_multi, [11, 13], return_rejected=True
+    )
+
+    selected = selected.read()
+    rejected = rejected.read()
+
+    assert list(selected.index) == [11, 13]
+    assert list(rejected.index) == [10, 12, 14]
+
+
+@pytest.mark.parametrize("TextFileReader", [False, True])
+def test_split_by_column_entries_basic(sample_df, sample_reader, TextFileReader):
+    if TextFileReader:
+        data = sample_reader
+    else:
+        data = sample_df
+
+    selected, rejected, _, _ = split_by_column_entries(
+        data, {"B": ["y"]}, return_rejected=True
+    )
+
+    if TextFileReader:
+        selected = selected.read()
+        rejected = rejected.read()
+
+    assert list(selected.index) == [11, 14]
+    assert list(rejected.index) == [10, 12, 13]
+
+
+@pytest.mark.parametrize(
+    "inverse, reset_index, exp_selected_idx, exp_rejected_idx",
+    [
+        (False, False, [], [10, 11, 12, 13, 14]),
+        (False, True, [], [10, 11, 12, 13, 14]),
+        (True, False, [10, 11, 12, 13, 14], []),
+        (True, True, [10, 11, 12, 13, 14], []),
+    ],
+)
+@pytest.mark.parametrize("chunked", [False, True])
+def test_split_by_boolean_basic_true(
+    sample_df,
+    sample_reader,
+    boolean_mask,
+    inverse,
+    reset_index,
+    exp_selected_idx,
+    exp_rejected_idx,
+    chunked,
+):
+    if chunked:
+        data = sample_reader
+    else:
+        data = sample_df
+
+    selected, rejected, _, _ = split_by_boolean(
+        data,
+        boolean_mask,
+        boolean=True,
+        inverse=inverse,
+        reset_index=reset_index,
+        return_rejected=True,
+    )
+
+    exp_selected = sample_df.loc[exp_selected_idx]
+    exp_rejected = sample_df.loc[exp_rejected_idx]
+
+    if reset_index is True:
+        exp_selected = exp_selected.reset_index(drop=True)
+        exp_rejected = exp_rejected.reset_index(drop=True)
+
+    if chunked:
+        selected = selected.read()
+        rejected = rejected.read()
+
+    pd.testing.assert_frame_equal(selected, exp_selected)
+    pd.testing.assert_frame_equal(rejected, exp_rejected)
+
+
+@pytest.mark.parametrize(
+    "inverse, reset_index, exp_selected_idx, exp_rejected_idx",
+    [
+        (False, False, [], [10, 11, 12, 13, 14]),
+        (False, True, [], [10, 11, 12, 13, 14]),
+        (True, False, [10, 11, 12, 13, 14], []),
+        (True, True, [10, 11, 12, 13, 14], []),
+    ],
+)
+@pytest.mark.parametrize("chunked", [False, True])
+def test_split_by_boolean_basic_false(
+    sample_df,
+    sample_reader,
+    boolean_mask,
+    inverse,
+    reset_index,
+    exp_selected_idx,
+    exp_rejected_idx,
+    chunked,
+):
+    if chunked:
+        data = sample_reader
+    else:
+        data = sample_df
+
+    selected, rejected, _, _ = split_by_boolean(
+        data,
+        boolean_mask,
+        boolean=False,
+        inverse=inverse,
+        reset_index=reset_index,
+        return_rejected=True,
+    )
+
+    exp_selected = sample_df.loc[exp_selected_idx]
+    exp_rejected = sample_df.loc[exp_rejected_idx]
+
+    if reset_index is True:
+        exp_selected = exp_selected.reset_index(drop=True)
+        exp_rejected = exp_rejected.reset_index(drop=True)
+
+    if chunked:
+        selected = selected.read()
+        rejected = rejected.read()
+
+    pd.testing.assert_frame_equal(selected, exp_selected)
+    pd.testing.assert_frame_equal(rejected, exp_rejected)
+
+
+@pytest.mark.parametrize("TextFileReader", [False, True])
+def test_split_by_boolean_true_basic(
+    sample_df, sample_reader, boolean_mask_true, TextFileReader
+):
+    if TextFileReader:
+        data = sample_reader
+    else:
+        data = sample_df
+
+    selected, rejected, _, _ = split_by_boolean_true(
+        data, boolean_mask_true, return_rejected=True
+    )
+
+    if TextFileReader:
+        selected = selected.read()
+        rejected = rejected.read()
+
+    assert list(selected.index) == [10]
+    assert list(rejected.index) == [11, 12, 13, 14]
+
+
+@pytest.mark.parametrize("TextFileReader", [False, True])
+def test_split_by_boolean_false_basic(
+    sample_df, sample_reader, boolean_mask, TextFileReader
+):
+    if TextFileReader:
+        data = sample_reader
+    else:
+        data = sample_df
+
+    selected, rejected, _, _ = split_by_boolean_false(
+        data, boolean_mask, return_rejected=True
+    )
+
+    if TextFileReader:
+        selected = selected.read()
+        rejected = rejected.read()
+
+    assert list(selected.index) == []
+    assert list(rejected.index) == [10, 11, 12, 13, 14]
+
+
+@pytest.mark.parametrize("TextFileReader", [False, True])
+def test_split_by_index_empty(empty_df, empty_reader, TextFileReader):
+    if TextFileReader:
+        data = empty_reader
+    else:
+        data = empty_df
+
+    selected, rejected, _, _ = split_by_index(data, [0, 1], return_rejected=True)
+
+    if TextFileReader:
+        selected = selected.read()
+        rejected = rejected.read()
+
+    assert selected.empty
+    assert rejected.empty
+
+
+@pytest.mark.parametrize("TextFileReader", [False, True])
+def test_split_by_column_empty(empty_df, empty_reader, TextFileReader):
+    if TextFileReader:
+        data = empty_reader
+    else:
+        data = empty_df
+
+    selected, rejected, _, _ = split_by_column_entries(
+        data, {"A": [1]}, return_rejected=True
+    )
+
+    if TextFileReader:
+        selected = selected.read()
+        rejected = rejected.read()
+
+    assert selected.empty
+    assert rejected.empty
+
+
+@pytest.mark.parametrize("TextFileReader", [False, True])
+def test_split_by_boolean_empty(empty_df, empty_reader, TextFileReader):
+    if TextFileReader:
+        data = empty_reader
+    else:
+        data = empty_df
+
+    mask = empty_df.astype(bool)
+    selected, rejected, _, _ = split_by_boolean(
+        data, mask, boolean=True, return_rejected=True
+    )
+
+    if TextFileReader:
+        selected = selected.read()
+        rejected = rejected.read()
+
+    assert selected.empty
+    assert rejected.empty
